refactor(runtime): inline javafx runtime lookup in launcher
Remove the dedicated runtime locator class and keep launcher bootstrap
resolution in one place to simplify startup flow maintenance.
